src/specialized-tools/requirements.txt
azure-functions>=1.18.0
azure-cosmos>=4.5.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

